Use IllegalStateException and IllegalArgumentException

+++ src/main/java/org/scify/jedai/blockprocessing/AbstractBlockProcessing.java
@@ -31,7 +31,7 @@ public abstract class AbstractBlockProcessing implements IBlockProcessing {
     
     protected void printOriginalStatistics(List<AbstractBlock> inputBlocks) {
         long comparisons = 0;
-        comparisons = inputBlocks.stream().map((block) -> block.getNoOfComparisons()).reduce(comparisons, (accumulator, _item) -> accumulator + _item);
+        comparisons = inputBlocks.stream().map((block) -> (long) block.getNoOfComparisons()).reduce(comparisons, (accumulator, _item) -> accumulator + _item);
         
         Log.info("Original blocks\t:\t" + inputBlocks.size());
         Log.info("Original comparisons\t:\t" + comparisons);
