+++ src/main/java/org/scify/jedai/datamodel/AbstractBlock.java
@@ -31,7 +31,7 @@ public abstract class AbstractBlock implements Serializable {
     
     protected int blockIndex;
     
-    protected long comparisons;
+    protected int comparisons;
     protected float entropy;
     protected float utilityMeasure;
             
@@ -53,7 +53,7 @@ public abstract class AbstractBlock implements Serializable {
         return entropy;
     }
 
-    public long getNoOfComparisons() {
+    public int getNoOfComparisons() {
         return comparisons;
     }
